ai/requirements.txt
# Core Data Processing
polars>=0.19.0
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & Deep Learning
torch>=2.0.0
pytorch-forecasting>=1.0.0
pytorch-lightning>=2.0.0

# Financial Data
yfinance>=0.2.28

# Vector Search & Similarity
faiss-cpu>=1.7.4

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
huggingface-hub>=0.17.0

# Logging & Warnings
warnings-stdlib
